Back/requirements.txt
django==4.2
djangorestframework
PyJWT
djangorestframework-simplejwt
psycopg2-binary  # Для PostgreSQL
mysqlclient       # Для MySQL
pymongo           # Для MongoDB
dnspython         # В некоторых случаях нужно для Mongo
celery
redis
pydantic
marshmallow
nltk
spacy
drf-yasg
sentry-sdk        # Для интеграции Sentry (логирование)
djongo
PyPDF2
python-docx
django-cors-headers